Impose python package structure
- aidistillery/ holds the actual package, for now, it is mainly data
  cleaning and other utilities
- our executable scripts reside in bin/
