+++ bin/embed_word2vec
@@ -1,3 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+
 """
 embed.py (Executable)
 
@@ -8,8 +12,9 @@ Example: `python3 embed.py -f data/text -o data/saved_embedding -d 100`
 
 import argparse
 from gensim.models import word2vec
-from utils.clean_raw_data import list_of_strings_to_list_of_lists, clean_raw_text_from_file, phrasing_sentences
-from utils import file_handling
+
+from aidistillery import file_handling
+from aidistillery.data_cleaning import list_of_strings_to_list_of_lists, clean_raw_text_from_file, phrasing_sentences
 
 
 def load_and_process(file_name, min_length):
